updaated all to be fp64

+++ raytracer-01.cpp
@@ -15,7 +15,7 @@ Implementation for https://gabrielgambetta.com/computer-graphics-from-scratch/02
 #include <math.h>
 #include <array>
 
-typedef std::array<float, 3> float3;
+typedef std::array<double, 3> double3;
 typedef std::array<uint8_t, 3> rgb;
 const rgb BACKGROUND_COLOR = {255, 255, 255};
 
@@ -50,25 +50,25 @@ bool put_pixel(
 // Linear Algebra
 
 // Compute dot product between two 3d vectors
-float dot(float3 v1, float3 v2) {
+double dot(double3 v1, double3 v2) {
     return v1[0] * v2[0] + v1[1] * v2[1] + v1[2] * v2[2];
 }
 
 // Elementwise subtraction between two 3d vectors
-float3 subtract(float3 a, float3 b) {
+double3 subtract(double3 a, double3 b) {
     return {a[0] - b[0], a[1] - b[1], a[2] - b[2]};
 }
 
 // Ray tracing
 
 struct Sphere {
-    float3 center;
-    float radius;
+    double3 center;
+    double radius;
     rgb color;
 
     Sphere() {}
 
-    Sphere(const float3& v_center, float v_radius, const rgb& v_color) {
+    Sphere(const double3& v_center, double v_radius, const rgb& v_color) {
         center = v_center;
         radius = v_radius;
         color = v_color;
@@ -76,24 +76,24 @@ struct Sphere {
 };
 
 // Convert 2d pixel coordinates to 3d viewport coordinates.
-float3 canvas_to_viewport(int32_t x, int32_t y, int32_t width, int32_t height) {
-    return { (float) x / width, (float) y / height, 1 };
+double3 canvas_to_viewport(int32_t x, int32_t y, int32_t width, int32_t height) {
+    return { (double) x / width, (double) y / height, 1 };
 }
 
 // Computes intersection of ray with spheres. Returns solutions in terms of
 // line parameter t.
-std::array<float, 2> intersect_ray_with_sphere(
-    float3 origin,
-    float3 direction,
+std::array<double, 2> intersect_ray_with_sphere(
+    double3 origin,
+    double3 direction,
     Sphere sphere
 ) {
-    float3 center = subtract(origin, sphere.center);
+    double3 center = subtract(origin, sphere.center);
 
-    float a = dot(direction, direction);
-    float b = 2 * dot(center, direction);
-    float c = dot(center, center) - sphere.radius * sphere.radius;
+    double a = dot(direction, direction);
+    double b = 2 * dot(center, direction);
+    double c = dot(center, center) - sphere.radius * sphere.radius;
 
-    float discriminant = b * b - 4 * a * c;
+    double discriminant = b * b - 4 * a * c;
     if (discriminant > 0) {
         return {
             (-b + sqrt(discriminant)) / (2 * a),
@@ -105,17 +105,17 @@ std::array<float, 2> intersect_ray_with_sphere(
 
 // Traces a ray against the spheres in the scene
 rgb trace_ray(
-    float3 origin,
-    float3 direction,
-    float min_t,
-    float max_t,
+    double3 origin,
+    double3 direction,
+    double min_t,
+    double max_t,
     std::vector<Sphere> spheres
 ) {
-    float closest_t = INFINITY;
+    double closest_t = INFINITY;
     Sphere closest_sphere;
 
     for (int i = 0; i < spheres.size(); i++) {
-        std::array<float, 2> ts = intersect_ray_with_sphere(origin, direction, spheres[i]);
+        std::array<double, 2> ts = intersect_ray_with_sphere(origin, direction, spheres[i]);
         if (ts[0] < closest_t && min_t < ts[0] && ts[0] < max_t) {
             closest_t = ts[0];
             closest_sphere = spheres[i];
@@ -140,7 +140,7 @@ int32_t main() {
     uint8_t data[width * height][3];
 
     // Define camera settings
-    float3 camera = {0, 0, 0};
+    double3 camera = {0, 0, 0};
 
     // Define scene
     std::vector<Sphere> spheres = {
@@ -152,7 +152,7 @@ int32_t main() {
     for (int32_t x = -width / 2; x < width / 2; x++) {
         for (int32_t y = -height / 2; y < height / 2; y++)
         {
-            float3 direction = canvas_to_viewport(x, y, width, height);
+            double3 direction = canvas_to_viewport(x, y, width, height);
             rgb color = trace_ray(camera, direction, 1.0f, INFINITY, spheres);
             put_pixel(data, width, height, x, y, color);
         }
