+++ raytracer-02.cpp
@@ -15,9 +15,9 @@ Implementation for https://gabrielgambetta.com/computer-graphics-from-scratch/03
 #include <math.h>
 #include <array>
 
-typedef std::array<float, 3> float3;
+typedef std::array<double, 3> double3;
 typedef std::array<uint8_t, 3> rgb;
-const float3 BACKGROUND_COLOR = {255, 255, 255};
+const double3 BACKGROUND_COLOR = {255, 255, 255};
 
 // Canvas
 
@@ -50,48 +50,48 @@ bool put_pixel(
 // Linear Algebra
 
 // Compute dot product between two 3d vectors
-float dot(float3 v1, float3 v2) {
+double dot(double3 v1, double3 v2) {
     return v1[0] * v2[0] + v1[1] * v2[1] + v1[2] * v2[2];
 }
 
 // Length of vector
-float length(float3 vec) {
+double length(double3 vec) {
     return sqrt(dot(vec, vec));
 }
 
 // Broadcasted multiply between scalar and a vector
-float3 multiply(float k, float3 vec) {
+double3 multiply(double k, double3 vec) {
     return {k * vec[0], k * vec[1], k * vec[2]};
 }
 
 // Elementwise addition between two 3d vectors
-float3 add(float3 a, float3 b) {
+double3 add(double3 a, double3 b) {
     return {a[0] + b[0], a[1] + b[1], a[2] + b[2]};
 }
 
 // Elementwise subtraction between two 3d vectors
-float3 subtract(float3 a, float3 b) {
+double3 subtract(double3 a, double3 b) {
     return {a[0] - b[0], a[1] - b[1], a[2] - b[2]};
 }
 
-rgb clamp(float3 vec) {
+rgb clamp(double3 vec) {
     return {
-        (uint8_t) std::round(std::clamp<float>(vec[0], 0.0f, 255.0)),
-        (uint8_t) std::round(std::clamp<float>(vec[1], 0.0f, 255.0)),
-        (uint8_t) std::round(std::clamp<float>(vec[2], 0.0f, 255.0)),
+        (uint8_t) std::round(std::clamp<double>(vec[0], 0.0f, 255.0)),
+        (uint8_t) std::round(std::clamp<double>(vec[1], 0.0f, 255.0)),
+        (uint8_t) std::round(std::clamp<double>(vec[2], 0.0f, 255.0)),
     };
 }
 
 // Ray tracing
 
 struct Sphere {
-    float3 center;
-    float radius;
-    float3 color;
+    double3 center;
+    double radius;
+    double3 color;
 
     Sphere() {}
 
-    Sphere(const float3& v_center, float v_radius, const float3& v_color) {
+    Sphere(const double3& v_center, double v_radius, const double3& v_color) {
         center = v_center;
         radius = v_radius;
         color = v_color;
@@ -102,12 +102,12 @@ enum LightType {AMBIENT, POINT, DIRECTIONAL};
 
 struct Light {
     LightType ltype;
-    float intensity;
-    float3 position;
+    double intensity;
+    double3 position;
 
     Light() {}
 
-    Light(LightType v_ltype, float v_intensity, const float3& v_position) {
+    Light(LightType v_ltype, double v_intensity, const double3& v_position) {
         ltype = v_ltype;
         intensity = v_intensity;
         position = v_position;
@@ -125,24 +125,24 @@ struct Scene {
 };
 
 // Convert 2d pixel coordinates to 3d viewport coordinates.
-float3 canvas_to_viewport(int32_t x, int32_t y, int32_t width, int32_t height) {
-    return { (float) x / width, (float) y / height, 1 };
+double3 canvas_to_viewport(int32_t x, int32_t y, int32_t width, int32_t height) {
+    return { (double) x / width, (double) y / height, 1 };
 }
 
 // Computes intersection of ray with spheres. Returns solutions in terms of
 // line parameter t.
-std::array<float, 2> intersect_ray_with_sphere(
-    float3 origin,
-    float3 direction,
+std::array<double, 2> intersect_ray_with_sphere(
+    double3 origin,
+    double3 direction,
     Sphere sphere
 ) {
-    float3 center = subtract(origin, sphere.center);
+    double3 center = subtract(origin, sphere.center);
 
-    float a = dot(direction, direction);
-    float b = 2 * dot(center, direction);
-    float c = dot(center, center) - sphere.radius * sphere.radius;
+    double a = dot(direction, direction);
+    double b = 2 * dot(center, direction);
+    double c = dot(center, center) - sphere.radius * sphere.radius;
 
-    float discriminant = b * b - 4 * a * c;
+    double discriminant = b * b - 4 * a * c;
     if (discriminant > 0) {
         return {
             (-b + sqrt(discriminant)) / (2 * a),
@@ -153,8 +153,8 @@ std::array<float, 2> intersect_ray_with_sphere(
 }
 
 // Compute lighting for the scene
-float compute_lighting(float3 point, float3 normal, std::vector<Light> lights) {
-    float intensity = 0;
+double compute_lighting(double3 point, double3 normal, std::vector<Light> lights) {
+    double intensity = 0;
     if (abs(length(normal) - 1.0f) > 0.0001f) {
         std::cerr << "Error: Normal is not length 1 (" << length(normal) << ")" << std::endl;
         return INFINITY;
@@ -165,14 +165,14 @@ float compute_lighting(float3 point, float3 normal, std::vector<Light> lights) {
         if (light.ltype == AMBIENT) {
             intensity += light.intensity;
         } else {
-            float3 vec_l;
+            double3 vec_l;
             if (light.ltype == POINT) {
                 vec_l = subtract(light.position, point);
             } else {  // Light.DIRECTIONAL
                 vec_l = light.position;
             }
 
-            float n_dot_l = dot(normal, vec_l);
+            double n_dot_l = dot(normal, vec_l);
             if (n_dot_l > 0) {
                 intensity += light.intensity * n_dot_l / (length(vec_l));
             }
@@ -183,18 +183,18 @@ float compute_lighting(float3 point, float3 normal, std::vector<Light> lights) {
 }
 
 // Traces a ray against the spheres in the scene
-float3 trace_ray(
-    float3 origin,
-    float3 direction,
-    float min_t,
-    float max_t,
+double3 trace_ray(
+    double3 origin,
+    double3 direction,
+    double min_t,
+    double max_t,
     Scene scene
 ) {
-    float closest_t = INFINITY;
+    double closest_t = INFINITY;
     Sphere closest_sphere;
 
     for (int i = 0; i < scene.spheres.size(); i++) {
-        std::array<float, 2> ts = intersect_ray_with_sphere(origin, direction, scene.spheres[i]);
+        std::array<double, 2> ts = intersect_ray_with_sphere(origin, direction, scene.spheres[i]);
         if (ts[0] < closest_t && min_t < ts[0] && ts[0] < max_t) {
             closest_t = ts[0];
             closest_sphere = scene.spheres[i];
@@ -209,11 +209,11 @@ float3 trace_ray(
         return BACKGROUND_COLOR;
     }
 
-    float3 point = add(origin, multiply(closest_t, direction));
-    float3 normal = subtract(point, closest_sphere.center);
+    double3 point = add(origin, multiply(closest_t, direction));
+    double3 normal = subtract(point, closest_sphere.center);
     normal = multiply(1.0f / length(normal), normal);
 
-    float intensity = compute_lighting(point, normal, scene.lights);
+    double intensity = compute_lighting(point, normal, scene.lights);
     return multiply(intensity, closest_sphere.color);
 }
 
@@ -224,7 +224,7 @@ int32_t main() {
     uint8_t data[width * height][3];
 
     // Define camera settings
-    float3 camera = {0, 0, 0};
+    double3 camera = {0, 0, 0};
 
     // Define scene
     std::vector<Sphere> spheres = {
@@ -243,8 +243,8 @@ int32_t main() {
     for (int32_t x = -width / 2; x < width / 2; x++) {
         for (int32_t y = -height / 2; y < height / 2; y++)
         {
-            float3 direction = canvas_to_viewport(x, y, width, height);
-            float3 color = trace_ray(camera, direction, 1.0f, INFINITY, scene);
+            double3 direction = canvas_to_viewport(x, y, width, height);
+            double3 color = trace_ray(camera, direction, 1.0f, INFINITY, scene);
             put_pixel(data, width, height, x, y, clamp(color));
         }
     }
